+++ src/test/java/io/github/astrapi69/swing/table/panel/shuffle/DemoPermissionsShuffleTablePanel.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.swing.table.shuffle.panel;
+package io.github.astrapi69.swing.table.panel.shuffle;
 
 import java.awt.*;
 import java.awt.event.ActionEvent;