+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 1.1-SNAPSHOT
 
 ADDED:
 
+- new TablePanel with a single table in a scroll pane
 - new methods for check if the table have single, multiply or no selection
 - all last relevant classes from swing-components repository
 - new method replaceWith that replaces the content with the given list of the GenericTableModel
@@ -16,6 +17,7 @@ CHANGED:
 - update of gradle-plugin dependency com.github.ben-manes.versions.gradle.plugin to new version 0.41.0
 - update of dependency swing-base-components to new version 1.1
 - update of dependency jobj-core to new version 5.2
+- moved all suffle panels to new package 'io.github.astrapi69.swing.table.panel.shuffle'
 
 Version 1
 -------------